Factory Demo Moscow campus
created and optimized a demo to be used for tours and showings for the new PLC course. This demo works in tandom with the Fa22 HMI.  The whole factory process is repeatable and allows full manual control with graceful ressetting of params. On the HMI you can select a puck to send through the factory, and can keep doing so for all other pucks. there is not internal state tracking in the sense of how many and which pucks were set already, so select ones that are available for now.