Updated From Moscow Fact
Pulled the latest program off the Moscow factory and replaced the old demo with it.  It seems like most of the rungs are the same, but in an effort to figure out whats wrong with the CDA factory, i needed to pull a known good copy. Old is the original demo file.